refactor: move channels_v2/slack_channel.py to channels/slack_channel.py
